refactor: consolidate libraries under new package
