+++ lib/__init__.py
@@ -1,10 +1,2 @@
-from .downloader5 import *           # includes your downloading logic
-from .make_clips import *            # includes clip-making routines
-from .tasks_lib import *             # core task logic
-from .transcription_utils import *   # transcription support
-from .video_utils import *           # logging, video functions
-from .watermarker2 import *          # watermarking logic
-from .whisper_utils import *         # Whisper-related functions
-from .url_utils import *             # URL helpers
-from .archive_utils import *         # archive download utilities
-
+"""Convenience imports for the mimesis toolkit."""
+from .mimesis import *
