+++ lib/mimesis/__init__.py
@@ -1,0 +1,17 @@
+"""Consolidated helper package.
+
+Modules are available as subpackages; no heavy
+imports occur at package import time.
+"""
+
+__all__ = [
+    "archive",
+    "clips",
+    "downloader",
+    "tasks",
+    "transcription",
+    "url",
+    "video",
+    "watermark",
+    "whisper",
+]
